+++ Packages/Sources/TopicFeature/Resources/Localizable.xcstrings
@@ -1,71 +1,14 @@
 {
   "sourceLanguage" : "en",
   "strings" : {
-    "Code" : {
-      "localizations" : {
-        "ru" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Код"
-          }
-        }
-      }
+    "[*]" : {
+
     },
-    "Code: %@" : {
-      "localizations" : {
-        "ru" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Код: %@"
-          }
-        }
-      }
+    "Код" : {
+
     },
-    "Quote" : {
-      "localizations" : {
-        "ru" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Цитата"
-          }
-        }
-      }
-    },
-    "Quote: %@" : {
-      "localizations" : {
-        "ru" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Цитата: %@"
-          }
-        }
-      }
-    },
-    "Quote: %@ @ %@" : {
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "new",
-            "value" : "Quote: %1$@ @ %2$@"
-          }
-        },
-        "ru" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Цитата: %1$@ @ %2$@"
-          }
-        }
-      }
-    },
-    "Whoops, something went wrong while parsing this post :(" : {
-      "localizations" : {
-        "ru" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Упс, что-то пошло не так во время загрузки этого поста :("
-          }
-        }
-      }
+    "Код: " : {
+
     },
     "Шапка Темы" : {
       "localizations" : {
